Add missing call to fclose() in runCmd()
This was causing us to exhaust available file handles.
One result was that API commands to return empty string after a while instead of the actual command result.
Another result was that we would be unable to reconnect to cgminer after a while (presumably the underlying socket code needed a file handle and could not allocate one).