Use the most recent data entry to determine which data keys to draw Areas for
NOTE: This should be much more reliable, however, it could be that the latest entry may not have any data fields if communication was lost from apiserver to cgminer.
This is just something to keep an eye our for.  Could probably be fixed by making a union of all keys across all data points.